Bug fix: Custom file actions display their Finder icon again (this includes custom AppleScript actions).
Note: This is more of a workaround since the check if an action's title looks like a path now happens second
whereas before this check was last. This means that whatever was tested second before might now not have icons
if the title of those actions may look like a path. In practice though I have not notice any more icon disappearance.