Remove entries concerning sdp conversion of sdef in Scripting.
This is done because there has been a report of someone having trouble with the sdp
not installed on the build system. I see no harm in keeping the generated files in there
as they haven't changed for ages and will help people with troubled build systems to 
not mess up at the scripting target.