Fix to error messages
 * Unhandled AE conversion from descriptor (null)
 * error {\n    NSAppleScriptErrorNumber = -1750;\n}

The first one is triggered if there are no arguments passed to
runAppleScript and so

returnDesc = [script executeAppleEvent:event error:&errorInfo];

is never run. One open topic is that this will also return NIL if there
is an error. Maybe handle this in the future.

The second error is displayed if there is no script behind the action
command. In this case default to the old behaviour of returning 1.