Initial loop optimizations
First refactoring of loops to use Obj-C 2.0 
"for...in" fast Enumeration feature.

Also, refactored all loops that used the "foreach"
macro to a non-macro version.  For most of these
the variables are now declared with strong types
when I was able to determine the type.
(the macro version declared the variables of type
id).