Issue 39 (Bug 262) Incorrect display of Websearch
This is a work around for websearches that display the wrong icon (multiple icons overlayed and looking messy). The real problem probably is within the Websearch Plugin itself, but I can't get breakpoints to trigger in it.

PS, just looked at http://coelomic.wordpress.com/2006/01/02/quicksilver-tips/
and see the google icon in image 4.  While fixing the initial problem I noticed that that spot was a blank square, and didn't know what, why, or how, it was like that, so I just overrode it to look like image 5, as a temporary fix. Wish I'd seen it sooner because I don't remember where the override is (probably in QSObject_URLHandling.m).