Refactored QSObject_URLHandling drawing code
There are 3 different drawing methods based on conditional compiling.

#define USE_ORIGINAL_URL_ICON_DRAWING_CODE 1
Uses the original drawing code that would overlay a white square for websearches.

#define USE_NEW_URL_ICON_DRAWING_CODE 1
The default drawing mode. This version does not do any custom drawing. It's able to work because the web search and Firefox objects get proper NSImages in QSResourceManager.m

#define USE_NEW_URL_ICON_DRAWING_CODE_WITH_FAVICONS 1
Experimental, based on the original drawing code. If a favicon image is matched to a URL then the favicon is also drawn drawn in the first panel.  Favicons are provided by Safari (will need the updated plugin as well).  This functiion still needs work as it does not find matches when the URL string starts with "qss-" or "qssp-".