Fix issues 26, 47, 57 and 61
Issue 26, enabled Commands tab in QSTriggersPrefPane.xib. See QSTriggersPrefPane.m selectedTrigger for notes about a UI change that should be made as well. I was unable to find where to make the change so it is left as a future modification. Otherwise it does not effect usability.

Issue 47, fixed. See comments in revised code for details.

Issue 57 and 61, fixed. Added activated state flag to QSTrigger and separating it from the enabled flag's logic. See QSTrigger.m function activated for more details.