Fix a crash when installing plug-ins.
This crash may be caused by any QSTask being dealloc'd, but it only seemed to happen when installing plug-ins. The bug seemed to be that -[QSTask stopTask] was being called, which removed self from tasksDictionary, which called -release and then -dealloc, which called -setIdentifier: with nil, which tried to remove self from tasksDictionary again, causing an EXC_BAD_ACCESS. The fix is to simply release identifier and clear it, since we know that's what it's going to do anyway.