Fix a bug in plugin loading when multiple versions of a plugin were available locally.
Now the system checks the local plugins instead of only plugins about to be loaded. The difference is that plugins that have dependencies don't get added to the about to be loaded list, and thus the first one found gets loaded (which might not be the latest).

Also, log an error in case an old plugin can't be removed.