Attempt to fix more multi-threading crashes
In QSCatalogEntry, calling [self contents] was itself calling `saveIndex`... causing recursive calls. Also, the `copy` was a vain attempt at fixing the problem. It only performs a shallow copy, leaving the NSArray's contents (NSMutableDicts to be precise) free to be mutated.
The QSLibrarian Fix ensures that catalog contents cannot be altered whilst they're being added to.