Improve the way matched strings are displayed in the UI
Previously, if the name (and not the label) matched, only the name would display. This could be ugly for e.g. long URLs and horribly named items (where the label was used to improve the object's appearance)

Now, the `displayName` is ALWAYS shown, whether it is matched or not. If the matched string is not the `displayName`, the matched string is shown on the line below