Attempt to display more useful iObjects for 3rd pane actions
These changes kick in if an action provider doesn't (correctly) implement the `validIndirectObjectsForAction...` method.
If this method returns nil, these changes attempt to get a list of objects based on the action's `indirectTypes` array.

This will not incredibly well at the moment, since QSLib cannot deal with UTIs, but the `indirectFileTypes` array can