allow for OS-specific resource definitions
In `ResourceLocations.plist` or a plug-in's `QSResourceAdditions `, you
can now specify different icons for different versions of OS X.

For example, if a resource with the key "Printer" exists on 10.6 and
10.7, but not 10.8, you could define two resources named
"Printer[10.8]" and "Printer". The first will only be used on 10.8
systems. All other OS versions will fall back to the latter resource.

It is also necessary (and possible) in rare circumstances to be more
specific, like "RemoteHost[10.7.3]".