simplify validation using the new infoRecord-based methods
This can be a bit slower than the old way if a lot of file objects don't have infoRecord cached, but it's not terrible (about 0.1s), and once the info is cached, it's about 10 times faster than the previous code.

This is a proof-of-concept. There are probably many other validation methods we could simplify.