Serial catalog entry scanning
This is a general rework of the changes already mentioned in #1232.
Serial dispatch queues have now been introduced to each Catalog entry. In essence, this means that each catalog entry is responsible for scanning in a serial (single thread) manner, as opposed to trying to get QSLibrarian to do it.