consolidate `setIcon` and `updateIcon`
The previous commit starts icon loading on another thread, but for the
UI to know when the icon is ready, every instance of
`loadIconForObject` would have to be modified to call `updateIcon`.
Almost all of them currently call `setIcon`. This should allow
`setIcon` to correctly handle updated icons, but without sending a
notification every time it's called.