remove the `previewIcon` method
For file objects, `loadIconForObject` was doing two things: loading a
more generic icon that can be gotten to quickly, then calling
`previewIcon` to handle the more expensive preview generation in the
background. The first task is the job of `setQuickIconForObject`, so
that code has been relocated to that method. The expensive preview
generation was moved from `previewIcon` to `loadIconForObject`, since
it always runs on a background thread now.