show the file's icon before checking for a Quick Look preview
There can be a noticeable delay when trying to get a preview, only to find there isn't one and the file's icon should be used instead. So, show that first.

If there is no QL preview, calling `setIcon:` a second time with the file's icon is a no-op, so it should be fine.

fixes #1552