Use public.utf8-plain-text instead of NSStringPboardType
The main place this affects is pasteboard objects and an object's 'stringValue'. Previously, the public.utf8-plain-text type was stored in a QSObject's dataDictionary as data, now it is converted to a string (along with any other string type).
When getting an object's stringValue, we must remember to look at the NSStringPboard type for backwards compatibility.