avoid an infinite stringValue loop
splitObjects tries three things: check if count is 1, check for object components, get children. We already know the count is not one, and getting children (if it doesn't cause a loop) just splits the string on \n, which we would only reverse here. So go straight for the object components.

fixes #2242