Another go at ensuring search string is cleared when closing interface
The problem is that clearing the search string too early means mnemonics are not saved, so if we move the mnemonic saving code earlier this fixes it.
The reason we want to clear the search string on interface close is a UI thing. See this comment: https://github.com/quicksilver/Quicksilver/pull/2249\#issuecomment-236476810
Fixes #2302
Fixes #2300
Related to #2268 #2269