Fix uploading of `ChangesBare.html`
`ChangesBare.html` was being created in the `build` job but not included
in the artifacts used for the `sign` job, which runs for new releases
and needs access to `ChangesBare.html` to update the changelog at
qsapp.com.

This ensures that ChangesBare.html is included in the dmg artifacts,
which has everything needed for signing and releasing a new version.

Fixes https://github.com/quicksilver/Quicksilver/issues/2958