Copy ChangesBare to dmg folder
This folder is created and uploaded during the build / step tests for
every push, but then only downloaded / signed for releases.

CI was failing to find `ChangesBare.html` to update the changelog
because it is created during the build steps but wasn't being added
to the second-stage artifacts (as it is not strictly needed for signing
/ distribution).

This simply adds it to those artifacts so it is availble for pushing
changelogs when we cut a release.