Don't automatically index private directories when testing
By default, `QSCorePlugIn-Info.plist` is configured to index directories
that newer editions of MacOS consider "sensitive," such as:
- `~/Desktop`
- `~/Downloads`
- `~/Documents`

Accessing these directories *immediately* provokes a permissions popup
that **blocks** until dismissed or accepted.

Unfortunately, this is problematic for testing (and especially headless
testing / CI) in that the tests block -- indefinitely -- until this
popup is acted on, whether or not these directories are actually
required for the test.

Thankfully none of our existing tests necessarily require access to any
of these directories. This commit provides an alternative file
(`QSCorePlugIn-Info-Testing.plist`) to be used only during the Testing
scheme that does *not* index these directories and thereby avoids this
problematic default. One test was changed to as to not access
`~/Downloads`, which seemed unnecessary for the function being tested.

Additionally, this commit sets the `QSDisableExternalPlugIns`
environment variable -- only for the Testing scheme -- which prevents
automatic loading of plugins in `~/Library/Application
Support/Quicksilver/PlugIns` (among other locations). For
reproducibility, these user-installed plugins should probably *not* be
automatically included during testing. For example, having the
`AddressBook` plugin enabled will trigger a permissions popup (similar
to those described above) at launch time, blocking the test
indefinitely, even though it is not required for the Quicksilver test
suite.