Remove the private APIs entirely
This removes `CGSPrivate.h`, `CPSPrivate.h`, as well as
`QSCGSTransition.{h,m}` which only seemed to serve as adapters for them.
It additionally either removes entirely or comments out (to serve as a
quick and dirty reference for reimplementation) code that depended on
these files.

This will break the pretty animations.

See also: https://github.com/quicksilver/Quicksilver/issues/3001